feat: add lazy mobile/desktop app components
