+++ apps/codeimage/src/App.tsx
@@ -1,29 +1,24 @@
-import {Frame} from './components/Frame/Frame';
-import {Canvas} from './components/Scaffold/Canvas/Canvas';
 import {Scaffold} from './components/Scaffold/Scaffold';
-import {CustomEditor} from './components/CustomEditor/CustomEditor';
-import {Toolbar} from './components/Toolbar/Toolbar';
 import {Sidebar} from './components/Scaffold/Sidebar/Sidebar';
-import {createEffect, createMemo, createSignal, on, Show} from 'solid-js';
-import {ThemeSwitcher} from './components/ThemeSwitcher/ThemeSwitcher';
-import {useFrameState} from './state/frame';
-import {useTerminalState} from './state/terminal';
-import {DynamicTerminal} from './components/Terminal/dynamic/DynamicTerminal';
-import {Footer} from './components/Footer/Footer';
+import {
+  createEffect,
+  createMemo,
+  lazy,
+  Match,
+  on,
+  Suspense,
+  Switch,
+} from 'solid-js';
 import {useUIState} from './state/ui';
 import {useI18n} from '@codeimage/locale';
 import {useModality} from './core/hooks/isMobile';
-import {BottomBar} from './components/BottomBar/BottomBar';
-import {FrameHandler} from './components/Frame/FrameHandler';
-import {EditorSidebar} from './components/LeftSidebar/EditorSidebar';
 import {NotificationHandler} from './components/ui/Toast/SnackbarHost';
 import ReloadPrompt from './components/PromptUpdate/PromptUpdate';
 
+const LazyMobileApp = lazy(() => import('./_MobileApp'));
+const LazyDesktopApp = lazy(() => import('./_DesktopApp'));
+
 const App = () => {
-  const [frameRef, setFrameRef] = createSignal<HTMLElement>();
-  const [portalHostRef, setPortalHostRef] = createSignal<HTMLElement>();
-  const frame = useFrameState();
-  const terminal = useTerminalState();
   const ui = useUIState();
   const modality = useModality();
   const [, {locale}] = useI18n();
@@ -35,63 +30,16 @@ const App = () => {
     <Scaffold>
       <NotificationHandler />
       <ReloadPrompt />
-      <Show when={modality === 'full'}>
-        <Sidebar>
-          <EditorSidebar />
-        </Sidebar>
-      </Show>
-
-      <div
-        ref={setPortalHostRef}
-        id={'portal-host'}
-        style={{
-          position: 'relative',
-          width: '0px',
-          height: '0px',
-          // eslint-disable-next-line solid/style-prop
-          'z-index': 10,
-        }}
-      />
-
-      <Canvas>
-        <Toolbar canvasRef={frameRef()} />
-
-        <FrameHandler ref={setFrameRef} onScaleChange={frame.setScale}>
-          <Frame
-            radius={0}
-            padding={frame.padding}
-            background={frame.background}
-            opacity={frame.opacity}
-            visible={frame.visible}
-          >
-            <DynamicTerminal
-              type={terminal.type}
-              readonlyTab={false}
-              tabName={terminal.tabName}
-              showTab={true}
-              shadow={terminal.shadow}
-              background={terminal.background}
-              accentVisible={terminal.accentVisible}
-              darkMode={terminal.darkMode}
-              textColor={terminal.textColor}
-              onTabChange={terminal.setTabName}
-              showHeader={terminal.showHeader}
-            >
-              <CustomEditor />
-            </DynamicTerminal>
-          </Frame>
-        </FrameHandler>
-
-        <Footer />
-      </Canvas>
-
-      {modality === 'mobile' ? (
-        <BottomBar portalHostRef={portalHostRef()} />
-      ) : (
-        <Sidebar>
-          <ThemeSwitcher orientation={'vertical'} />
-        </Sidebar>
-      )}
+      <Suspense fallback={'Loading...'}>
+        <Switch>
+          <Match when={modality === 'mobile'}>
+            <LazyMobileApp />
+          </Match>
+          <Match when={modality === 'full'}>
+            <LazyDesktopApp />
+          </Match>
+        </Switch>
+      </Suspense>
     </Scaffold>
   );
 };
