+++ apps/codeimage/src/components/LeftSidebar/EditorSidebar.tsx
@@ -2,8 +2,9 @@ import {EditorForm} from './EditorForm';
 import {WindowStyleForm} from './WindowStyleForm';
 import {FrameStyleForm} from './FrameStyleForm';
 import {EditorStyleForm} from './EditorStyleForm';
+import {JSXElement} from 'solid-js';
 
-export const EditorSidebar = () => {
+export default function EditorSidebar(): JSXElement {
   return (
     <EditorForm>
       <FrameStyleForm />
@@ -13,4 +14,4 @@ export const EditorSidebar = () => {
       <EditorStyleForm />
     </EditorForm>
   );
-};
+}
